Revelation

Matrix Types and functions

Types

data Channel

Promoted data type to track channel type at the type level.

Constructors

RGB

BGR

Grayscale

HSV

YUV

data Mat c elem

Matrix type - type parameter c must have kind Channel. Use extract to get at the underlying C'Mat ptr.

data InverseMethod

The decomposition method used to compute the inverse

Constructors

LU

Cholesky

SVD

invert :: Mat c e -> CV (Mat c e)

Inverts the given matrix. Uses SVD to guarantee that *something* is always returned. The matrix provided is guaranteed to at least be a pseudo-inverse (left or right). Gives a true inverse when the matrix is square and non-singular.

invertBy :: InverseMethod -> Mat c e -> CV (Mat c e)

Inverts the given matrix with the provided method. This function is partial if the method passed is not SVD *and* a singular or non-square matrix is passed in.

type family ElemT c :: * -> *

Type synonym family to deal with the storage format for matrices of various channels. If you need to pass something of this type to a constructor, it's always safe to pass (undefined :: ElemT c e) with c and e filled in as appropriate.

class CVElement f where

Typeclass to convert the provided element type to an OpenCV element type.

Methods

cvElemType :: f -> CInt

Constructors

createMat :: CVElement (ElemT c e) => Int -> Int -> ElemT c e -> CV (Mat c e)

A safe matrix creation function. This one gives you exactly what you ask for, and guarantees that a matrix of the appropriate type and size is allocated. The ElemT c e parameter is safely passed as undefined because it's value is never inspected. It's needed as a type witness. The witness can't be created automatically with GHC 7.6 because type families are open and not guaranteed to be injective.

unsafeCreateMat :: CV (Mat c e)

This return type potentially lies... use only if you're truly polymorphic in the type of the underlying matrix.

createIdentity :: CVElement (ElemT c e) => Int -> ElemT c e -> CV (Mat c e)

Allocates a square identity matrix of size r x r

ones :: CVElement (ElemT c e) => Int -> Int -> ElemT c e -> CV (Mat c e)

Allocates a matrix of the requested size and fills it with ones.

zeros :: CVElement (ElemT c e) => Int -> Int -> ElemT c e -> CV (Mat c e)

Allocates a matrix of the requested size and fills it with zeroes.

Functions

rows :: Integral a => Mat c e -> CV a

Extract the number of rows in the provide matrix (equivalent to field accessor rows on the C++ side)

cols :: Integral a => Mat c e -> CV a

Extract the number of rows in the provide matrix (equivalent to field accessor cols on the C++ side)

index :: Storable (ElemT c e) => Mat c e -> V2 Int -> CV (ElemT c e)

Index into a matrix at (row, column) given by (V2 row column).

fromMat :: Storable (ElemT c e) => Mat c e -> CV (Vector (Vector (ElemT c e)))

Create a Vector (of Storable.Vectors) (provided for integration with linear)

transpose :: Mat c e -> CV (Mat c e)

Undocumented

newtype CV a

Constructors

CV

Fields

runCV :: IO a

liftCV :: MonadTrans m => IO a -> m CV a

cameraCapture :: Int -> VideoCapture c e

type VideoCapture c e = Producer (Mat c e) CV ()

type Window c e = Consumer (Mat c e) CV ()

imageDisplayWindow :: String -> Window c e

data Channel

Promoted data type to track channel type at the type level.

Constructors

RGB

BGR

Grayscale

HSV

YUV

convertColorP :: Convertable c c' => Pipe (Mat c e) (Mat c' e) CV ()

convertColor :: Convertable c c' => Mat c e -> CV (Mat c' e)

waitKey :: Integral a => a -> CV Char

waitKeyP :: Integral a => Maybe Char -> a -> Pipe (Mat c e) (Mat c e) CV ()

indexP :: Pipe (Mat Grayscale Int) (Mat Grayscale Int) CV ()